+++ src/app/components/timeline-mini.tsx
@@ -1,0 +1,37 @@
+import { FC , useRef } from 'react';
+import GlitchText from './glitch-text';
+
+function TimelineNode({ children , title , marginTop=0 } : { children : JSX.Element , title? : string , marginTop : number }) {
+    return (
+        <>
+        <div className="w-[40px] h-full flex flex-col items-center relative">
+            <div className="h-full w-[4px] bg-primary-500" />
+        </div>
+        <div className={`w-full h-auto min-h-screen transition-all mt-[${marginTop}px]`}>
+            <div className='flex items-center'>
+                <div className="absolute translate-x-[-30px] w-[20px] h-[20px] rounded-full bg-primary-500" />
+                { title ? <GlitchText className='text-8xl font-bold text-secondary-200'>{ title }</GlitchText> : null }
+            </div>
+            { children }
+        </div>
+        </>
+    )
+}
+
+interface TimelineItem {
+    children : JSX.Element[];
+    className?: string;
+    titles?: string[];
+}
+
+const Timeline: FC<TimelineItem> = ({ children , className , titles }) => {
+    let timelineNodeList = children.map((child, index) => <TimelineNode title={ titles && titles[index] } marginTop={index == 0 ? 100 : 0}>{ child }</TimelineNode>);
+
+    return (
+        <div className={`grid grid-cols-[auto_1fr] ${ className }`}>
+            { timelineNodeList }
+        </div>
+    )
+}
+
+export default Timeline;
