+++ src/app/sections/announcements.tsx
@@ -1,0 +1,5 @@
+export default function Announcements() {
+    return (
+        <div className=""></div>
+    )
+}
